Move functions, required in types.py to _util.py
Add __all__ to util.py for sphinx to generate docs properly
